+++ packages/polaris-viz/src/components/DonutChart/components/InnerValue/InnerValue.tsx
@@ -1,4 +1,4 @@
-import {Fragment} from 'react';
+import {Fragment, useEffect, useState} from 'react';
 import {useSpring, animated, config} from '@react-spring/web';
 import type {LabelFormatter} from '@shopify/polaris-viz-core';
 import {useTheme} from '@shopify/polaris-viz-core';
@@ -9,12 +9,15 @@ import type {ComparisonMetricProps} from '../../../ComparisonMetric';
 import {ComparisonMetric} from '../../../ComparisonMetric';
 import styles from '../../DonutChart.scss';
 
+const SCALING_FACTOR = 0.07;
+
 export interface InnerValueProps {
   activeValue: number | null | undefined;
   activeIndex: number;
   labelFormatter: LabelFormatter;
   isAnimated: boolean;
   totalValue: number;
+  diameter: number;
   comparisonMetric?: ComparisonMetricProps;
   renderInnerValueContent?: RenderInnerValueContent;
 }
@@ -27,8 +30,10 @@ export function InnerValue({
   isAnimated,
   renderInnerValueContent,
   totalValue,
+  diameter,
 }: InnerValueProps) {
   const selectedTheme = useTheme();
+  const [fontSize, setFontSize] = useState(0);
 
   const {animatedValue} = useSpring({
     animatedValue: totalValue,
@@ -38,6 +43,10 @@ export function InnerValue({
       immediate: !isAnimated,
     },
   });
+
+  useEffect(() => {
+    setFontSize(diameter * SCALING_FACTOR);
+  }, [diameter]);
 
   const animatedTotalValue = (
     <animated.span>
@@ -63,7 +72,7 @@ export function InnerValue({
     <Fragment>
       <animated.p
         className={classNames(styles.ContentValue)}
-        style={{color: selectedTheme.xAxis.labelColor}}
+        style={{color: selectedTheme.xAxis.labelColor, fontSize}}
       >
         {valueToDisplay}
       </animated.p>